TL;DR - moved all files and directories to the root
in order to use this repo as a submodule in other "child" repos, it needs to be in a dedicated directory at the root of the calling repo. In order then for wordpress to see it as a theme, it will need to be a sibling directory of the child theme contained in the calling repo.