moving functionality to determine page title in here.
originally, Content class handled determining a page's title.  This information would be handed down to Header class to determine header title and used in the final views.  But since we are now instantiating the Header class BEFORE we call Content, it's too late for Header to use it.  Moving that functionality into Main so we can call it before we instantiate Header.